Resolve DB SNP RS id value
ColocatedVariants in the Genome Nexus response can return non-dbSNP ids, such as COSMIC ids. A regex is used to match for colocated variant ids that match the dbSNP convention (rs###). The first dbSNP id encountered is the value that will be used for populating the "dbSNP_RS" field for the record.

- Adjusted test output files for both isoforms (mskcc, uniprot)

Signed-off-by: Angelica Ochoa <15623749+ao508@users.noreply.github.com>