Handle changes to deps in cbioportal v3.6.4
Somehow the old logging configuration won't work anymore (getting some
logconfiguration error). This fixes that by reverting some changes to
loggin framework exclusion. In addition the the spring boot version was
upgraded. Peculiarly some in memory database had to be included. Not
entirely sure how the dependency tree works, but it is probably due to
changes in the databaseAnnotator configuration in the new spring boot
version.